Fixing limitations with using systems with multiple system instances
This change introduces a global limitation for the concept of a System Instance. As Ansible uses the environment as a unique key to retrieve the relevant System Instance, only one System Instance for each environment can be used. This is not an acceptable solution in the longer term and will be addressed.

Ansible completely fails in terms of managing results from an API, i.e. selecting an index and then re-indexing the array so which index was matched is in index 0. Instead Ansible will leave the index position unchanged, meaning we have no reliable way to determine which index to use. To workaround this we request hte System Instance directly from the API, rather than get all Instances and determine which one to use.