Refreshing facts before looking up host information
This is to cover situations where the system-hostname role is applied to a host in the same play as this role.

Ansible will load facts before the system-hostname is updated and have the old value cached when this role looks up host information by FQDN.

This fix forces facts to be refreshed so it is usable in the same play.